+++ app/src/app/[locale]/page.tsx
@@ -1,52 +1,16 @@
-'use client';
-
+import LocaleButton from '~/components/LocaleButton';
+import TestAPI from '~/components/TestAPI';
 import { env } from '~/env.ts';
-import { useChangeLocale, useCurrentLocale, useI18n } from '~/locales/client';
-import type { LocaleCode } from '~/locales/server';
-import { api } from '~/server/clients/client';
-import { Suspense } from 'react';
+import { getI18n } from '~/locales/server';
+import { setStaticParamsLocale } from 'next-international/server';
 
-function IndexPage() {
-  const hello = api.hello.hello.useQuery(
-    { text: 'client' },
-    {
-      refetchInterval: 1000,
-    },
-  );
-  if (!hello.data) {
-    return <div>Loadddding...</div>;
-  }
-  return (
-    <div>
-      <p>{hello.data.greeting}</p>
-    </div>
-  );
-}
-
-function LocaleButton({ locale }: { locale: LocaleCode }) {
-  const activeLocale = useCurrentLocale();
-  const isActive = locale === activeLocale;
-  const changeLocale = useChangeLocale({ preserveSearchParams: true });
-
-  return (
-    <Suspense>
-      <button
-        onClick={() => changeLocale(locale)}
-        disabled={isActive}
-        className="m-3 rounded border-2 border-red-500 p-2 disabled:cursor-not-allowed disabled:opacity-50"
-      >
-        {locale}
-      </button>
-    </Suspense>
-  );
-}
-
-export default function Bidule() {
-  const t = useI18n();
+export default async function Bidule({ params }: { params: { locale: string } }) {
+  setStaticParamsLocale(params.locale);
+  const t = await getI18n();
 
   return (
     <main className="yeye flex min-h-screen flex-col items-center justify-between p-24">
-      {/* <IndexPage /> */}
+      <TestAPI />
       <p>{env.NEXT_PUBLIC_TESTVALUE}</p>
       <p>bidule</p>
       <p>rebidule</p>
